requirements.txt
Django==1.8.5
django-registration-redux==1.2
djangorestframework
docxtpl
simplejson
django-crispy-forms